+++ src/screens/LazyLoader/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Page from 'containers/LazyLoader/page';
+import Page from 'screens/LazyLoader/page';
 
 const LazyLoader = () => {
   return <Page />;
