+++ src/screens/LazyLoader/loader.js
