+++ src/screens/Login/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Page from 'containers/Login/page';
+import Page from 'screens/Login/page';
 
 const Login = () => {
   return <Page />;
