+++ src/tests/screens/ForgotPassword/page.test.js
@@ -2,11 +2,12 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import Page from 'containers/Login/page';
 import backgroundImage from 'images/auth_background.png';
+import Page from 'screens/ForgotPassword/page';
 
 describe('When visited the Login Page', () => {
-  const logoLabel = 'Sign in to Nimble';
+  const logoLabel =
+    'Enter your email to receive instructions for resetting your password.';
   it('shows a background image', () => {
     const { getByLabelText } = render(<Page />);
     const imageDiv = getByLabelText('background-layout-image');
