+++ src/tests/screens/Home/index.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import Home from 'containers/Home';
+import Home from 'screens/Home';
 
 describe('When visited the forgot password page', () => {
   it('renders page name Forgot Password', () => {
