+++ src/tests/screens/Login/page.test.js
@@ -2,12 +2,11 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import Page from 'containers/ForgotPassword/page';
 import backgroundImage from 'images/auth_background.png';
+import Page from 'screens/Login/page';
 
 describe('When visited the Login Page', () => {
-  const logoLabel =
-    'Enter your email to receive instructions for resetting your password.';
+  const logoLabel = 'Sign in to Nimble';
   it('shows a background image', () => {
     const { getByLabelText } = render(<Page />);
     const imageDiv = getByLabelText('background-layout-image');
